GCP_cloud_functions/raw_to_data_lake/requirements.txt
ndjson
google-cloud-storage